House, Ambulance, and Hospital IDs: We’ll add IDs for each house, ambulance, and hospital and display these IDs on the canvas.
Ambulance Colors Based on States:

Red: Ambulance is on its way to pick up a patient.
Yellow: Ambulance has picked up the patient and is heading to the hospital.
Green: Ambulance is free and available to pick up a patient.
House Colors:

Red: House has a patient waiting for pickup.
Green: House does not have a patient.
Ambulance Coordination:

If an ambulance is already on its way to a house, no other ambulances will go to that house. Other ambulances will either go to another house or remain idle at the hospital until a new patient request arises.